refactor observation display using hifitime, improve doc and add changelog for new version
